+++ src/aihwkit/simulator/rpu_base_src/rpu_base_tiles_cuda.cpp
@@ -51,39 +51,87 @@ void declare_rpu_tiles_cuda(py::module &m) {
           }),
           py::arg("cpu_tile"))
       .def(
+          "set_shared_weights",
+          [](Class &self, torch::Tensor weights) {
+            CHECK_TORCH_CUDA_INPUT(weights);
+            // weight is d major for CUDA !
+            if (weights.dim() != 2 || weights.size(1) != self.getDSize() ||
+                weights.size(0) != self.getXSize()) {
+              throw std::runtime_error(
+                  "Invalid weights dimensions: expected [" + std::to_string(self.getXSize()) + "," +
+                  std::to_string(self.getDSize()) + "] array");
+            }
+            if (weights.device().index() != self.getGPUId()) {
+              throw std::runtime_error(
+                  "Weights need to be on the same cuda device: expected " +
+                  std::to_string(self.getGPUId()) + ", got " +
+                  std::to_string(weights.device().index()));
+            }
+
+            std::lock_guard<std::mutex> lock(self.mutex_);
+            return self.setSharedWeights(weights.data_ptr<T>());
+          },
+          py::arg("weights"))
+      .def(
+          "set_delta_weights",
+          [](Class &self, torch::Tensor delta_weights) {
+            CHECK_TORCH_CUDA_INPUT(delta_weights);
+            // weight is d major for CUDA !
+            if (delta_weights.dim() != 2 || delta_weights.size(1) != self.getDSize() ||
+                delta_weights.size(0) != self.getXSize()) {
+              throw std::runtime_error(
+                  "Invalid delta weights dimensions: expected [" + std::to_string(self.getXSize()) +
+                  "," + std::to_string(self.getDSize()) + "] array");
+            }
+            if (delta_weights.device().index() != self.getGPUId()) {
+              throw std::runtime_error(
+                  "Delta weights need to be on the same cuda device: expected " +
+                  std::to_string(self.getGPUId()) + ", got " +
+                  std::to_string(delta_weights.device().index()));
+            }
+
+            std::lock_guard<std::mutex> lock(self.mutex_);
+            return self.setDeltaWeights(delta_weights.data_ptr<T>());
+          },
+          py::arg("delta_weights"))
+
+      .def(
           "forward",
           [](Class &self, const torch::Tensor &x_input_, bool bias = false, bool x_trans = false,
              bool d_trans = false, bool is_test = false) {
             auto x_input = x_input_.contiguous();
             CHECK_TORCH_CUDA_INPUT(x_input);
 
-            if (x_input.dim() != 2) {
-              throw std::runtime_error("Invalid x_input dimensions: expected 2 dimensional array");
+            if (x_input.dim() < 1) {
+              throw std::runtime_error(
+                  "Invalid x_input dimensions: expected at least 1 dimensional array");
             }
-
-            int in_size = x_trans ? x_input.size(0) : x_input.size(1);
+            int in_size = x_trans ? x_input.size(0) : x_input.size(-1);
             int expected_in_size = self.getXSize() - (bias ? 1 : 0);
-            int m_batch = x_trans ? x_input.size(1) : x_input.size(0);
+            int m_batch = x_input.numel() / in_size;
             int out_size = self.getDSize();
 
             // Validate the x_input dimensions.
             if (in_size != expected_in_size) {
-              std::string shape_str = x_trans ? ("[_, " + std::to_string(expected_in_size) + "]")
-                                              : ("[" + std::to_string(expected_in_size) + ",_]");
+              std::string shape_str = x_trans ? ("[*, " + std::to_string(expected_in_size) + "]")
+                                              : ("[" + std::to_string(expected_in_size) + ",*]");
               throw std::runtime_error(
                   "Invalid x_input dimensions: expected " + shape_str + " array");
             }
 
+            // Build the buffers.
+            std::vector<int64_t> dims(x_input.sizes().begin(), x_input.sizes().end());
+            if (d_trans) {
+              dims[0] = out_size;
+            } else {
+              dims[dims.size() - 1] = out_size;
+            }
+            torch::Tensor d_output = torch::empty(dims, x_input.options());
+
+            // Call RPU function.
             self.finishUpdateCalculations();
             std::lock_guard<std::mutex> lock(self.mutex_);
             self.setStream(at::cuda::getCurrentCUDAStream()); // TODO: better way to get the stream?
-
-            // Build the buffers.
-            int dim0 = d_trans ? out_size : m_batch;
-            int dim1 = d_trans ? m_batch : out_size;
-            torch::Tensor d_output = torch::empty({dim0, dim1}, x_input.options());
-
-            // Call RPU function.
             self.forward(
                 x_input.template data_ptr<T>(), d_output.template data_ptr<T>(), bias, m_batch,
                 x_trans, d_trans, is_test);
@@ -95,27 +143,28 @@ void declare_rpu_tiles_cuda(py::module &m) {
           R"pbdoc(
            Compute the dot product (forward pass).
 
-           Compute the dot product::
+           Compute the dot product:
+           .. math:
 
-               Y = X * W (+ b)
+               \mathbf{y} = W\mathbf{x} [+ \mathbf{b}]
 
-           where ``X`` is the input and ``W`` is the ``[d_size, x_size]``
+           where :math:`\mathbf{x}` is the input and :math:`W` is the ``[d_size, x_size]``
            current weight matrix. If ``bias`` is True, then it is assumes that a
-           bias row is added to the analog tile weights.  The input ``X`` is
+           bias row is added to the analog tile weights.  The input :math:`\mathbf{x}` is
            then  expected to be of size ``x_size -1`` , as internally it will be
            expanded by a 1, to match the bias row in the tile weights.
 
            An analog tile will have a possible non-ideal version of this forward pass.
 
            Args:
-               x_input: ``[N, x_size (- 1)]`` torch::tensor.
+               x_input: ``[N,*, x_size (- 1)]`` input :math:`\mathbf{x}` torch::Tensor.
                bias: whether to use bias.
-               x_trans: whether the ``x_input`` matrix is transposed.
+               x_trans: whether the ``x_input`` matrix is transposed. That is of size ``[x_size (- 1), *, N]``
                d_trans: whether the ``d`` matrix is transposed.
                is_test: whether inference (true) mode or training (false)
 
            Returns:
-               torch::tensor: ``[N, d_size]`` torch::tensor.
+               torch::tensor: ``[N, *, d_size]`` or ``[d_size, *, N]`` matrix.
            )pbdoc")
 
       .def(
@@ -125,33 +174,36 @@ void declare_rpu_tiles_cuda(py::module &m) {
             auto d_input = d_input_.contiguous();
             CHECK_TORCH_CUDA_INPUT(d_input);
 
-            if (d_input.dim() != 2) {
-              throw std::runtime_error("Invalid d_input dimensions: expected 2 dimensional array");
+            if (d_input.dim() < 1) {
+              throw std::runtime_error(
+                  "Invalid d_input dimensions: expected at least 1 dimensional array");
             }
-
-            int in_size = d_trans ? d_input.size(0) : d_input.size(1);
+            int in_size = d_trans ? d_input.size(0) : d_input.size(-1);
             int expected_in_size = self.getDSize();
-            int m_batch = d_trans ? d_input.size(1) : d_input.size(0);
+            int m_batch = d_input.numel() / in_size;
             int out_size = self.getXSize() - (bias ? 1 : 0);
 
             // Validate the d_input dimensions.
             if (in_size != expected_in_size) {
-              std::string shape_str = d_trans ? ("[_, " + std::to_string(expected_in_size) + "]")
-                                              : ("[" + std::to_string(expected_in_size) + ",_]");
+              std::string shape_str = d_trans ? ("[*, " + std::to_string(expected_in_size) + "]")
+                                              : ("[" + std::to_string(expected_in_size) + ", *]");
               throw std::runtime_error(
                   "Invalid d_input dimensions: expected " + shape_str + " array");
             }
 
+            // Build the buffers.
+            std::vector<int64_t> dims(d_input.sizes().begin(), d_input.sizes().end());
+            if (x_trans) {
+              dims[0] = out_size;
+            } else {
+              dims[dims.size() - 1] = out_size;
+            }
+            torch::Tensor x_output = torch::empty(dims, d_input.options());
+
+            // Call RPU function.
             self.finishUpdateCalculations();
             std::lock_guard<std::mutex> lock(self.mutex_);
             self.setStream(at::cuda::getCurrentCUDAStream());
-
-            // Build the buffers.
-            int dim0 = x_trans ? out_size : m_batch;
-            int dim1 = x_trans ? m_batch : out_size;
-            torch::Tensor x_output = torch::empty({dim0, dim1}, d_input.options());
-
-            // Call RPU function.
             self.backward(
                 d_input.template data_ptr<T>(), x_output.template data_ptr<T>(), bias, m_batch,
                 d_trans, x_trans);
@@ -162,23 +214,24 @@ void declare_rpu_tiles_cuda(py::module &m) {
           R"pbdoc(
            Compute the transposed dot product (backward pass).
 
-           Compute the transposed dot product::
+           Compute the transposed dot product:
+           .. math:
 
-               Y = D * W'
+               \mathbf{y} = W\mathbf{d}
 
-           where ``D`` is the input and ``W'`` is the ``[d_size, x_size]`` transposed current
-           weight matrix.
+           where :math:`\mathbf{d}` is the input and :math:`W` is the  current
+           weight matrix (of size ``[d_size, x_size]``).
 
            An analog tile will have a possible non-ideal version of this backward pass.
 
            Args:
-               d_input: ``[N, d_size]`` torch::tensor.
+               d_input: ``[N, *,  d_size]`` input :math:`\mathbf{d}` torch::Tensor.
                bias: whether to use bias.
-               x_trans: whether the ``x_input`` matrix is transposed.
-               d_trans: whether the ``d`` matrix is transposed.
+               d_trans: whether the ``d_input`` matrix is transposed. That is of size ``[d_size, *, N]``
+               x_trans: whether the ``x`` output matrix is transposed.
 
            Returns:
-               torch::tensor: ``[N, x_size (-1)]`` torch::tensor.
+               torch::Tensor: ``[N, *, x_size (-1)]`` or ``[x_size (-1), *, N]`` torch::Tensor.
            )pbdoc")
 
       .def(
@@ -191,30 +244,30 @@ void declare_rpu_tiles_cuda(py::module &m) {
             CHECK_TORCH_CUDA_INPUT(d_input);
             CHECK_TORCH_CUDA_INPUT(x_input);
 
-            if ((x_input.dim() != 2) || (d_input.dim() != 2)) {
+            if ((x_input.dim() < 1) || (d_input.dim() < 1)) {
               throw std::runtime_error(
-                  "Invalid x_input/d_input dimensions: expected 2 dimensional array");
+                  "Invalid x_input/d_input dimensions: expected at least 1 dimensional array");
             }
 
-            int in_size = x_trans ? x_input.size(0) : x_input.size(1);
+            int in_size = x_trans ? x_input.size(0) : x_input.size(-1);
             int expected_in_size = self.getXSize() - (bias ? 1 : 0);
-            int m_batch = x_trans ? x_input.size(1) : x_input.size(0);
+            int m_batch = x_input.numel() / in_size;
 
-            int out_size = d_trans ? d_input.size(0) : d_input.size(1);
+            int out_size = d_trans ? d_input.size(0) : d_input.size(-1);
             int expected_out_size = self.getDSize();
-            int m_batch_from_d = d_trans ? d_input.size(1) : d_input.size(0);
+            int m_batch_from_d = d_input.numel() / out_size;
 
             // Validate the x_input dimensions.
             if (in_size != expected_in_size) {
-              std::string shape_str = x_trans ? ("[_, " + std::to_string(expected_in_size) + "]")
-                                              : ("[" + std::to_string(expected_in_size) + ",_]");
+              std::string shape_str = x_trans ? ("[*, " + std::to_string(expected_in_size) + "]")
+                                              : ("[" + std::to_string(expected_in_size) + ", *]");
               throw std::runtime_error(
                   "Invalid x_input dimensions: expected " + shape_str + " array");
             }
             // Validate the d_input dimensions.
             if (out_size != expected_out_size) {
-              std::string shape_str = d_trans ? ("[_, " + std::to_string(expected_out_size) + "]")
-                                              : ("[" + std::to_string(expected_out_size) + ",_]");
+              std::string shape_str = d_trans ? ("[*, " + std::to_string(expected_out_size) + "]")
+                                              : ("[" + std::to_string(expected_out_size) + ", *]");
               throw std::runtime_error(
                   "Invalid d_input dimensions: expected " + shape_str + " array");
             }
@@ -238,11 +291,12 @@ void declare_rpu_tiles_cuda(py::module &m) {
           R"pbdoc(
            Compute an n-rank update.
 
-           Compute an n-rank update::
+           Compute an n-rank update:
+           .. math:
 
-               W += -LR * D * X'
+               W \leftarrow W - \lambda \mathbf{x}\mathbf{d}^T
 
-           where ``LR`` is the learning rate.
+           where :math:`\lambda` is the learning rate.
 
            An analog tile will have a possible non-ideal version of this update pass.
 
@@ -250,11 +304,11 @@ void declare_rpu_tiles_cuda(py::module &m) {
                The learning rate is always positive, and thus scaling is negative.
 
            Args:
-               x_input: ``[N, x_size (-1)]`` torch::tensor.
-               d_input: ``[N, d_size]`` torch::tensor.
+               x_input: ``[N, *, x_size (-1)]`` input :math:`\mathbf{x}` torch::Tensor.
+               d_input: ``[N, *, d_size]`` input :math:`\mathbf{d}` torch::Tensor.
                bias: whether to use bias.
-               x_trans: whether the ``x_input`` matrix is transposed.
-               d_trans: whether the ``d`` matrix is transposed.
+               x_trans: whether the ``x_input`` matrix is transposed, ie. ``[x_size (-1), *, N]``
+               d_trans: whether the ``d`` matrix is transposed, ie. ``[d_size, *, N]``
            )pbdoc")
       .def(
           "forward_indexed",